refactor: InputDevice 패키지 변경(domain -> view)
