+++ src/main/java/lotto/view/InputDevice.java
@@ -1,4 +1,4 @@
-package lotto.domain;
+package lotto.view;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.Arrays;
